fix(access): pin linspace dtype for CI's numpy stubs
CI's Python 3.10 numpy types np.linspace as floating[Any], which mypy --strict
rejects against the ArraySeconds annotation; the local stubs infer float64. An
explicit dtype satisfies both. CI had been red since the access merge - my own
rule says check it after every push, and I pushed two merges without doing so.

+++ src/orbital_engine/access.py
@@ -338,7 +338,7 @@ def access_grid(horizon_s: ScalarSeconds, sample_dt_s: ScalarSeconds) -> ArraySe
     if step <= 0.0:
         raise ValueError(f"sample_dt_s must be positive, got {step}.")
     n_intervals = max(1, int(round(horizon / step)))
-    grid: ArraySeconds = np.linspace(0.0, horizon, n_intervals + 1)
+    grid: ArraySeconds = np.linspace(0.0, horizon, n_intervals + 1, dtype=np.float64)
     return grid
